Audacity CommunityFOSSad
I took another approach making this CommunityFOSSad, the illustration is
taking all the screen size and illustrating a microphone and Audacity in
the background. The warm aspect of the pictures gives it a feel of ease
and comfort, I didn't want to interfere too much with the ambiance setup
with the illustration.